# api-server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Empreendedorismo",
    "description": "API para o projeto de empreendedorismo",
    "contact": {
      "email": "bbrezolin@hotmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://18.230.85.64",
      "description": "AWS Server"
    }
  ],
  "tags": [
    {
      "name": "users",
      "description": "Endpoints relacionados à usuários"
    },
    {
      "name": "trips",
      "description": "Endpoints relacionados à viagens"
    }
  ],
  "components": {
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "email",
          "password",
          "firstName",
          "lastName"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "email": {
            "type": "string",
            "example": "john.doe@email.com"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "Admin",
              "Manager",
              "Guide",
              "Tourist"
            ]
          },
          "firstName": {
            "type": "string",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "currentLocation": {
            "type": "string",
            "example": "{lat: 2.4234, long: 0.93298}"
          }
        }
      },
      "Trip": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Shopping in NYC"
          },
          "place": {
            "type": "string",
            "example": "5th avenue, NYC"
          },
          "startDate": {
            "type": "string",
            "example": "2023-04-23T19:15:12.600Z"
          },
          "endDate": {
            "type": "string",
            "example": "2023-04-23T19:16:12.600Z"
          },
          "isActivated": {
            "type": "boolean",
            "default": false
          },
          "areaCenter": {
            "type": "string",
            "example": "{lat: 2.4234, long: 0.93298}"
          },
          "areaRadius": {
            "type": "number",
            "format": "float64"
          },
          "meetingPoint": {
            "type": "string",
            "example": "{lat: 2.4234, long: 0.93298}"
          },
          "meetingTime": {
            "type": "string",
            "example": "2023-04-23T19:15:12.600Z"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Mensagem de Erro."
          }
        }
      }
    }
  },
  "paths": {
    "/users": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Cria um novo usuário",
        "operationId": "createUser",
        "requestBody": {
          "description": "Usuário a ser criado",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "Faz login de um usuário",
        "operationId": "loginUser",
        "requestBody": {
          "description": "Usuário a ser logado",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "john.doe@email.com",
                    "description": "Email do usuário"
                  },
                  "password": {
                    "type": "string",
                    "example": "password",
                    "description": "Senha do usuário"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário logado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "User": {
                      "$ref": "#/components/schemas/User"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/trips": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Cria uma nova viagem",
        "operationId": "createTrip",
        "security": [
          {
            "jwt": [
              "Manager"
            ]
          }
        ],
        "requestBody": {
          "description": "Viagem a ser criada",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trip"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Viagem criada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/trips/{id}": {
      "get": {
        "tags": [
          "trips"
        ],
        "summary": "Retorna uma viagem",
        "operationId": "getTrip",
        "security": [
          {
            "jwt": [
              "Tourist"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Viagem retornada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Trip"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Erro interno do servidor",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "trips/{id}/area": {
      "post": {
        "tags": [
          "trips"
        ],
        "summary": "Configura a área segura de uma viagem",
        "operationId": "addAreaToTrip",
        "security": [
          {
            "jwt": [
              "Manager"
            ]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID da viagem",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Área a ser adicionada",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "radius": {
                    "type": "number",
                    "format": "float"
                  },
                  "area": {
                    "type": "object",
                    "properties": {
                      "lat": {
                        "type": "number"
                      },
                      "long": {
                        "type": "number"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "responses": {
        "200": {
          "description": "Área adicionada com sucesso",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Trip"
              }
            }
          }
        },
        "400": {
          "description": "Erro de validação",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Error"
              }
            }
          }
        },
        "500": {
          "description": "Erro interno do servidor",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Error"
              }
            }
          }
        }
      }
    }
  }
}